{"schema":"decodex/commit/1","summary":"Split Decodex app operator readout components","authority":"XY-1115"}
