+++ apps/decodex-app/Sources/DecodexApp/OperatorLaneReadoutSupport.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+func rawPanelToken(_ value: String) -> String {
+	value.trimmingCharacters(in: .whitespacesAndNewlines)
+}
